### 2020-02-08 #2087 @oliviaguest
Hi @pmcharrison, if you have anybody in mind that you think is especially suitable please let me know. I have already sent off some invitations but I value your input too, of course — and thank you for submitting to JOSS. 😺
