### 2020-02-10 #2088 @whedon
[REVIEW]: psychTestR: An R package for designing and conducting behavioural psychological experiments
**Submitting author:** @pmcharrison (<a href="http://orcid.org/0000-0002-9851-9462">Peter Harrison</a>)
**Repository:** <a href="https://github.com/pmcharrison/psychTestR" target ="_blank">https://github.com/pmcharrison/psychTestR</a>
**Version:** v2.19.0
