### 2020-02-10 #2088 @oliviaguest
@Cognitive-Modeller, @DivyaSeernani, @pmcharrison thank you all again for submitting/accepting to review this. Start ideally by reading the checklist at the very top of this page and then looking at the code and documentation. :blush:

If you need to discuss things that are very specific to the code of this submission feel free to open an issue in the [code's repo itself](https://github.com/pmcharrison/psychTestR) and then link back to it from here.

Thank you!  :sparkles:
