### 2020-02-11 #2087 @pmcharrison
> Hi @pmcharrison, if you have anybody in mind that you think is especially suitable please let me know. I have already sent off some invitations but I value your input too, of course — and thank you for submitting to JOSS.

Hi @oliviaguest, thanks very much for finding reviewers so quickly! I was not sure who to suggest myself, but thank you for offering the opportunity.
