### 2020-02-26 #2088 @DivyaSeernani
Hia, just finished the review. Wasn't sure about the question on the license file since the .txt seems to contain a lot less information than the .md. Also there are clear community guidelines on the website, not in the readme in the repo. I guess that's ok? Otherwise, looks good and I'm looking forward to using the package for future experiments.
