### 2020-03-02 #2088 @pmcharrison
Hello @DivyaSeernani, thank you very much for this careful checking!

Regarding the community guidelines, I agree that it would be good to include information in the readme as well as on the website. I've therefore added the following section to the readme:

> The [psychTestR documentation website](https://pmcharrison.github.io/psychTestR/) answers many common questions about psychTestR usage. Further questions are welcomed via the [GitHub issue tracker](https://github.com/pmcharrison/psychTestR/issues). This issue tracker is also the place to report issues or software problems. We welcome software contributions in the form of [GitHub pull requests](https://github.com/pmcharrison/psychTestR/pulls); if you are thinking of making a contribution, consider first beginning a discussion on the issue tracker so that we can discuss implementation details.

Regarding the license file, yes, this reflects an awkward inconsistency between GitHub and CRAN license requirements (see https://stackoverflow.com/questions/43550479/how-to-satisfy-both-cran-and-github-license-file-naming-requirements and https://github.com/tidyverse/tidyr/issues/262 for discussions). Basically, CRAN doesn't want people to include full license texts, but GitHub does (and it's necessary to include the full license when distributing code). The current solution with the two files reflects a common solution implemented in the usethis package (see `usethis::use_mit_license`).
