### 2020-03-02 #2088 @oliviaguest
@Cognitive-Modeller just setting a reminder for you — Wednesday is better but if not this should provide an appropriate reminder for both of us. 😺
