### 2020-03-04 #2088 @oliviaguest
@Cognitive-Modeller @DivyaSeernani Thank you both for your reviews so far. I noticed that you both have left this one unchecked:

>  Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

Is this an ongoing issue and has @pmcharrison addressed this yet? :relaxed:
