### 2020-03-04 #2088 @pmcharrison
I guess there is the following in the article, which closest approximates a performance claim:

> The psychTestR server model is designed with scalability in mind, using an efficient single-process model that shares memory between sessions and hence easily scales to hundreds of simultaneous participants.

My best evidence for this claim is that we've run these kinds of highly parallel experiments on school students, and seen that scaling works well. A more quantitative test could be done using something like the [shinyloadtest framework](https://rstudio.github.io/shinyloadtest/) - let me know if you think I need to do something like this.
