### 2020-03-04 #2088 @oliviaguest
> My best evidence for this claim is that we've run these kinds of highly parallel experiments on school students, and seen that scaling works well. A more quantitative test could be done using something like the shinyloadtest framework - let me know if you think I need to do something like this.

@Cognitive-Modeller @DivyaSeernani I leave this up to you to resolve and discuss with @pmcharrison. 👍 When both reviewers have commented/ticked the box, we can continue to the next step. Thanks all for your contributions so far!
