### 2020-03-10 #2088 @pmcharrison
Hi @Cognitive-Modeller, @DivyaSeernani - what do you think about this performance claim question? Do you think I need to run and report some scaling tests, or do you think it's fine to rely on personal experience using psychTestR in studies?
