### 2020-03-10 #2088 @oliviaguest
@Cognitive-Modeller see @pmcharrison's comment aboive in relation to:

>  Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

:point_up: :relaxed:
