### 2020-04-02 8acd7b1
add style to trigger button
