+++ R/test-elements.R
@@ -529,6 +529,7 @@ describe_valid_p_id <- function() {
 NAFC_page <- function(label, prompt, choices, labels = NULL,
                       save_answer = TRUE,
                       arrange_vertically = length(choices) > 2L,
+                      style = "",
                       hide_response_ui = FALSE,
                       response_ui_id = "response_ui",
                       on_complete = NULL,
@@ -542,6 +543,7 @@ NAFC_page <- function(label, prompt, choices, labels = NULL,
                  labels = labels,
                  hide = hide_response_ui,
                  arrange_vertically = arrange_vertically,
+                 style = style,
                  id = response_ui_id))
   get_answer <- function(input, ...) input$last_btn_pressed
   validate <- function(answer, ...) !is.null(answer)
@@ -576,7 +578,7 @@ NAFC_page <- function(label, prompt, choices, labels = NULL,
 #' @export
 make_ui_NAFC <- function(choices, labels = NULL, hide = FALSE,
                          arrange_vertically = length(choices) > 2L,
-                         id = "response_ui") {
+                         style = "", id = "response_ui") {
   stopifnot(is.character(choices), length(choices) > 0L, is.scalar.logical(hide),
             is.null(labels) ||
               ((is.character(labels) || is.list(labels)) &&
@@ -587,7 +589,7 @@ make_ui_NAFC <- function(choices, labels = NULL, hide = FALSE,
   shiny::tags$div(id = id,
                   style = if (hide) "visibility: hidden" else "visibility: inherit",
                   mapply(function(id, label) {
-                    trigger_button(inputId = id, label = label)
+                    trigger_button(inputId = id, label = label, style = style)
                   }, choices, labels, SIMPLIFY = F, USE.NAMES = F) %>%
                     (function(x) if (arrange_vertically) lapply(x, shiny::tags$p) else x))
 }
@@ -884,7 +886,7 @@ dropdown_page.get_answer <- function(alternative_text) {
 #'
 #' @export
 trigger_button <- function(inputId, label, icon = NULL, width = NULL,
-                           enable_after = 0,
+                           enable_after = 0, style = "",
                            ...) {
   checkmate::qassert(enable_after, "N1[0,)")
   inputId <- htmltools::htmlEscape(inputId, attribute = TRUE)
@@ -894,6 +896,7 @@ trigger_button <- function(inputId, label, icon = NULL, width = NULL,
       icon = icon, width = width,
       onclick = "trigger_button(this.id);",
       disabled = TRUE,
+      style = style,
       ...),
     shiny::tags$script(
       sprintf("setTimeout(function() {
