### 2020-04-04 bc27fe1
add dummy version info to login page
