### 2020-04-04 65a6243
add app_info to footer
