+++ R/options.R
@@ -109,7 +109,8 @@ test_options <- function(title, admin_password,
                          logo_height = NULL,
                          display = display_options(),
                          allow_url_rewrite = TRUE,
-                         advance_delay = 0) {
+                         advance_delay = 0,
+                         app_info = "") {
   stopifnot(is.character(title),
             is.scalar.character(admin_password),
             is.null.or(researcher_email, is.scalar.character),
@@ -139,7 +140,8 @@ test_options <- function(title, admin_password,
               (is.scalar.character(logo_width) && is.scalar.character(logo_height)),
             is.list(display),
             is.scalar.logical(allow_url_rewrite),
-            is.scalar.numeric(advance_delay))
+            is.scalar.numeric(advance_delay),
+            is.character(app_info))
   # if (is.null(session_dir)) session_dir <- get_default_session_dir()
 
   if (!allow_url_rewrite && enable_resume_session) {
@@ -232,6 +234,7 @@ test_options <- function(title, admin_password,
        logo_height = logo_height,
        display = display,
        allow_url_rewrite = allow_url_rewrite,
+       app_info = app_info,
        js_opt = list(advance_delay = advance_delay))
 }
 
