### 2020-04-29 fc7d1bc
Adding an input check for conditional()
