+++ pkgdown.yml
@@ -1,5 +1,5 @@
 pandoc: '2.2'
-pkgdown: 1.4.1
+pkgdown: 1.5.1
 pkgdown_sha: ~
 articles:
   a1-simple-demo: a1-simple-demo.html
@@ -15,4 +15,5 @@ articles:
   d-url-parameters: d-url-parameters.html
   e-internationalised-questionnaire: e-internationalised-questionnaire.html
   e-jspsych-integration: e-jspsych-integration.html
+last_built: 2020-04-29T16:15Z
 
