### 2020-05-04 #2088 @pmcharrison
Hi @oliviaguest, @Cognitive-Modeller, @DivyaSeernani - just to add to the above, we had a live test today with 120 simultaneous participants and performance didn't suffer at all, I'm told. We've also had an online data collection over the last couple of months where we tested c. 300,000 participants (though not all at the same time!), and this seemed to work fine. Any chance we could move forward now?
