### 2020-05-04 #2088 @oliviaguest
> @Cognitive-Modeller see @pmcharrison's comment aboive in relation to:
>
> > Performance: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
>
> ☝️ ☺️

@pmcharrison I'm hoping @Cognitive-Modeller can report back. I'll send him an email just in case.
