### 2020-05-05 #2088 @arfon
> @openjournals/joss-eics please see above issue. I get 404 too. 🤔

This looks to be fixed now (you might have to reset your browser cache).

@Cognitive-Modeller, @DivyaSeernani - many thanks for your reviews here and to @oliviaguest for editing this one.
